refactor(types): move ListMaterializedViewsResponse to dbmcp-postgres
Materialized views are PostgreSQL-only, so the response type does not
need to live in the shared `dbmcp-server` crate. Move the struct, its
constructors, and the unit tests into `dbmcp-postgres::types`. Add
`indexmap` to the postgres crate's dependencies (already present in the
workspace).

`ListEntries` and the request type stay where they were — `ListEntries`
is reused by every list-tool response, and the request type was already
defined in the postgres crate. No wire change.
