+++ crates/postgres/src/types.rs
@@ -6,12 +6,14 @@
 //! can keep importing them from `crate::types`.
 
 use dbmcp_server::pagination::Cursor;
+use indexmap::IndexMap;
 use schemars::JsonSchema;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 pub use dbmcp_server::types::{
-    ListEntries, ListFunctionsResponse, ListMaterializedViewsResponse, ListProceduresResponse, ListTablesResponse,
-    ListTriggersRequest, ListTriggersResponse, ListViewsResponse,
+    ListEntries, ListFunctionsResponse, ListProceduresResponse, ListTablesResponse, ListTriggersRequest,
+    ListTriggersResponse, ListViewsResponse,
 };
 
 /// Request for the `dropTable` tool.
@@ -113,6 +115,37 @@ pub struct ListMaterializedViewsRequest {
     pub detailed: bool,
 }
 
+/// Response for the `listMaterializedViews` tool.
+#[derive(Debug, Serialize, JsonSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct ListMaterializedViewsResponse {
+    /// Page of matching materialized views. Shape depends on the request's `detailed` flag.
+    pub materialized_views: ListEntries,
+    /// Opaque cursor pointing to the next page. Absent when this is the final page.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub next_cursor: Option<Cursor>,
+}
+
+impl ListMaterializedViewsResponse {
+    /// Builds a brief-mode response from a page of bare matview names.
+    #[must_use]
+    pub fn brief(materialized_views: Vec<String>, next_cursor: Option<Cursor>) -> Self {
+        Self {
+            materialized_views: ListEntries::Brief(materialized_views),
+            next_cursor,
+        }
+    }
+
+    /// Builds a detailed-mode response from a page of name → metadata entries.
+    #[must_use]
+    pub fn detailed(materialized_views: IndexMap<String, Value>, next_cursor: Option<Cursor>) -> Self {
+        Self {
+            materialized_views: ListEntries::Detailed(materialized_views),
+            next_cursor,
+        }
+    }
+}
+
 /// Request for the Postgres `listProcedures` tool — extends the shared shape with `search` and `detailed`.
 #[derive(Debug, Default, Deserialize, JsonSchema)]
 #[serde(rename_all = "camelCase")]
@@ -136,8 +169,12 @@ pub struct ListProceduresRequest {
 
 #[cfg(test)]
 mod tests {
+    use indexmap::IndexMap;
+    use serde_json::json;
+
     use super::{
-        ListFunctionsRequest, ListMaterializedViewsRequest, ListProceduresRequest, ListTablesRequest, ListViewsRequest,
+        ListEntries, ListFunctionsRequest, ListMaterializedViewsRequest, ListMaterializedViewsResponse,
+        ListProceduresRequest, ListTablesRequest, ListViewsRequest,
     };
 
     #[test]
@@ -211,5 +248,28 @@ mod tests {
             serde_json::from_str(r#"{"search": "orders", "detailed": true}"#).expect("parse");
         assert_eq!(req.search.as_deref(), Some("orders"));
         assert!(req.detailed);
+    }
+
+    #[test]
+    fn list_materialized_views_response_brief_constructor_wraps_vec() {
+        let response = ListMaterializedViewsResponse::brief(vec!["mv_recent_orders".into()], None);
+        assert!(matches!(response.materialized_views, ListEntries::Brief(ref v) if v == &["mv_recent_orders"]));
+        assert!(response.next_cursor.is_none());
+    }
+
+    #[test]
+    fn list_materialized_views_response_detailed_constructor_wraps_indexmap() {
+        let map = IndexMap::from([("mv_recent_orders".into(), json!({"populated": true}))]);
+        let response = ListMaterializedViewsResponse::detailed(map, None);
+        assert!(matches!(response.materialized_views, ListEntries::Detailed(_)));
+    }
+
+    #[test]
+    fn list_materialized_views_response_brief_matches_legacy_wire_shape() {
+        let response = ListMaterializedViewsResponse::brief(vec!["mv_recent_orders".into()], None);
+        assert_eq!(
+            serde_json::to_value(&response).unwrap(),
+            json!({"materializedViews": ["mv_recent_orders"]})
+        );
     }
 }
