+++ crates/server/src/types.rs
@@ -302,37 +302,6 @@ impl ListProceduresResponse {
     }
 }
 
-/// Response for the `listMaterializedViews` tool.
-#[derive(Debug, Serialize, JsonSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ListMaterializedViewsResponse {
-    /// Page of matching materialized views. Shape depends on the request's `detailed` flag.
-    pub materialized_views: ListEntries,
-    /// Opaque cursor pointing to the next page. Absent when this is the final page.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub next_cursor: Option<Cursor>,
-}
-
-impl ListMaterializedViewsResponse {
-    /// Builds a brief-mode response from a page of bare matview names.
-    #[must_use]
-    pub fn brief(materialized_views: Vec<String>, next_cursor: Option<Cursor>) -> Self {
-        Self {
-            materialized_views: ListEntries::Brief(materialized_views),
-            next_cursor,
-        }
-    }
-
-    /// Builds a detailed-mode response from a page of name → metadata entries.
-    #[must_use]
-    pub fn detailed(materialized_views: IndexMap<String, Value>, next_cursor: Option<Cursor>) -> Self {
-        Self {
-            materialized_views: ListEntries::Detailed(materialized_views),
-            next_cursor,
-        }
-    }
-}
-
 /// Request for the `writeQuery` tool.
 #[derive(Debug, Default, Deserialize, JsonSchema)]
 #[serde(rename_all = "camelCase")]
@@ -589,29 +558,6 @@ mod tests {
         assert_eq!(
             serde_json::to_value(&response).unwrap(),
             json!({"views": ["active_users"]})
-        );
-    }
-
-    #[test]
-    fn list_materialized_views_response_brief_constructor_wraps_vec() {
-        let response = super::ListMaterializedViewsResponse::brief(vec!["mv_recent_orders".into()], None);
-        assert!(matches!(response.materialized_views, ListEntries::Brief(ref v) if v == &["mv_recent_orders"]));
-        assert!(response.next_cursor.is_none());
-    }
-
-    #[test]
-    fn list_materialized_views_response_detailed_constructor_wraps_indexmap() {
-        let map = IndexMap::from([("mv_recent_orders".into(), json!({"populated": true}))]);
-        let response = super::ListMaterializedViewsResponse::detailed(map, None);
-        assert!(matches!(response.materialized_views, ListEntries::Detailed(_)));
-    }
-
-    #[test]
-    fn list_materialized_views_response_brief_matches_legacy_wire_shape() {
-        let response = super::ListMaterializedViewsResponse::brief(vec!["mv_recent_orders".into()], None);
-        assert_eq!(
-            serde_json::to_value(&response).unwrap(),
-            json!({"materializedViews": ["mv_recent_orders"]})
         );
     }
 }
